requirements.txt
gradio==4.19.1
pandas==2.2.0
Pillow==10.2.0
Requests==2.31.0
transformers==4.36.2
torch==2.1.2